Happy path working, refactoring begins. \n# Move connect to own route
