+++ main.ts
@@ -1,13 +1,13 @@
 import Express from "express";
-import { createServer } from "http";
-import { Server } from "socket.io";
 import Cors from "cors";
 import morgan from "morgan";
-import winston from "winston";
 import { Client } from "irc-framework";
 import UserSettings from "./config";
 import MongooseDal from "./services/mongo";
 import { IChannel, IMessage } from "./models/channel";
+import { routes } from './router';
+import { createServer } from "http";
+import { Server } from "socket.io";
 
 MongooseDal.connect().then(
   () => {
@@ -20,92 +20,18 @@ MongooseDal.connect().then(
 let socketConnections = [];
 
 const app = Express();
+const httpServer = createServer(app);
 
 app.use(Cors());
 app.use(Express.json());
 app.use(Express.urlencoded({ extended: true }));
 app.use(morgan("dev"));
 
-const logger = winston.createLogger({
-  level: "info",
-  format: winston.format.json(),
-  defaultMeta: { service: "irc" },
-  transports: [
-    new winston.transports.Console(),
-  ],
-});
-
-const httpServer = createServer(app);
-const io = new Server(httpServer, {
-  cors: {
-    origin: "*",
-    methods: ["GET", "POST"],
-  },
-});
-
 const client = new Client();
 
+app.use('/', routes);
+
 app.use("/static", Express.static("public"));
-
-app.post("/connect", async (req, res) => {
-  if (!client.connected) client.connect(UserSettings);
-
-  var userInfo: IChannel = {
-    name: UserSettings.nick,
-    description: "Users NS Channel",
-    owner: UserSettings.nick,
-    created_at: new Date(),
-    updated_at: new Date(),
-    messages: [],
-    active: true
-  };
-
-  MongooseDal.createChannel(userInfo);
-
-  client.on("socket close", (e) => {
-    console.log(e);
-    console.log("socket closed");
-  });
-
-  client.on(
-    "message",
-    async (event: { nick: any; target: string; message: any }) => {
-      // TODO: This is for the POC - this should have a DTO and also is XSS vulnerable.
-      logger.info({
-        user: event.nick,
-        channel: event.target,
-        message: event.message,
-      });
-      io.emit("chat:message", {
-        user: event.nick,
-        channel: event.target,
-        message: event.message,
-      });
-
-      const messageStore: IMessage = {
-        sender: event.nick, // TODO : get session user
-        message: event.message,
-        created_at: new Date(),
-      };
-
-      console.log(messageStore);
-
-      await MongooseDal.addMessage(
-        Utils.CleanChannel(event.target),
-        messageStore
-      );
-    }
-  );
-
-  client.on("error", (event) => {
-    console.log(event);
-  });
-
-  // Get Users State from the mongoDB
-  var usersChannels = await MongooseDal.getChannelsForUser(UserSettings.nick);
-  res.send( {nick: UserSettings.nick, state: usersChannels} );
-
-});
 
 app.post("/channel/join", async (req, res) => {
   var channel = client.channel("#" + req.body.channel, req.body.key);
@@ -166,8 +92,14 @@ app.post('/nick/set', (req, res) => {
   res.send({message: "Nick changed", nick: req.body.nick });
 });
 
+// Move this to some service. 
+const io = new Server(httpServer, {
+  cors: {
+    origin: "*",
+    methods: ["GET", "POST"],
+  },
+});
 
-// Socket Hooks
 io.on("connection", (socket) => {
   console.log("a user connected : quasui Session ID: " + socket.id);
   // TODO: add by userID later
@@ -188,15 +120,6 @@ io.on("connection", (socket) => {
     await MongooseDal.addMessage(message.channel, messageStore);
   });
 });
-
-const Utils = {
-  CleanChannel: (channel: string) => {
-    return channel.replace("#", "");
-  },
-  FormChannel: (channel: string) => {
-    return "#" + channel;
-  },
-};
 
 httpServer.listen(3000, () => {
   console.log("listening on *:3000");
