+++ router/channel/index.ts
@@ -1,0 +1,3 @@
+import { Router } from "express";
+
+export const channelRouter = Router();
