+++ services/utils.ts
@@ -1,0 +1,10 @@
+const Utils = {
+    CleanChannel: (channel: string) => {
+      return channel.replace("#", "");
+    },
+    FormChannel: (channel: string) => {
+      return "#" + channel;
+    },
+  };
+
+export default Utils;
